feat: provided search term saved to the local storage & componentDidUpdate in CardList.tsx
