+++ src/components/MainSection.tsx
@@ -1,12 +1,12 @@
 import { Component } from 'react';
-import Card from './Card';
+import CardList from './CardList';
 
 class MainSection extends Component {
   render() {
     return (
       <main className="main" role="main">
         <div className="main-container">
-          <Card />
+          <CardList />
         </div>
       </main>
     );
